KVM: x86: Define RMP page fault error bits for #NPF
When SEV-SNP is enabled globally, the hardware places restrictions on
all memory accesses based on the RMP entry corresponding to the page
being accessed. When hardware encounters an RMP access violation during
a guest access, it will cause a #VMEXIT(NPF) with a number of additional
bits set to indicate the reasons for the #NPF. Define those bits here.

See APM2 section 16.36.10 for more details.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>
Signed-off-by: Ashish Kalra <ashish.kalra@amd.com>
[mdr: add some additional details to commit message]
Signed-off-by: Michael Roth <michael.roth@amd.com>