KVM: SEV: Add support to handle GHCB GPA register VMGEXIT
For SNP guests, before a vCPU uses a GHCB page the first time, it must
first register that GHCB page with the hypervisor via a special VMGEXIT
type defined by the GHCB specification. The hypervisor will then confirm
the GPA is valid/accessible and then keep track of it for subsequent
VMGEXIT handling. If a VMGEXIT is triggered while a guest's GHCB MSR
points to an unregistered GHCB GPA, the guest will be aborted.

Implement that handling here.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>
Signed-off-by: Ashish Kalra <ashish.kalra@amd.com>
Signed-off-by: Michael Roth <michael.roth@amd.com>