KVM: x86: Export the kvm_zap_gfn_range() for the SNP use
While resolving an RMP page fault, there may be cases where the page
level indicated by the RMP entry for the page is 2MB, whereas the TDP
needs to map the page as 4K in the nested page table. This can happen,
for example, if a guest previously PVALIDATED a 2MB private page, but
then later transitions one of the sub-pages to a shared page. In this
case the 2MB RMP entry needs to be split/PSMASH'd into individual 4K
entries, and the mapping level in the TDP needs to be adjusted
accordingly.

To handle this, the GFN range needs to be zapped after splitting the
pages in the RMP entry so it can get re-mapped with the new page level.
Export kvm_zap_gfn_range() so it can be used for this purpose.

Signed-off-by: Brijesh Singh <brijesh.singh@amd.com>
Signed-off-by: Ashish Kalra <ashish.kalra@amd.com>
Signed-off-by: Michael Roth <michael.roth@amd.com>