KVM: SVM: Make VMCB page allocation for nested SVM guest SNP safe
Implement a workaround for an SNP erratum where the CPU will incorrectly
signal an RMP violation #PF if a hugepage (2mb or 1gb) collides with the
RMP entry of a VMCB for a nested SVM guest.

Though vmcb12 page will be allocated by the L1 Hypervisor the vmcb that
actually will be setup on CPU for L1 Hypervisor will be setup by
L0 Hypervisor and that should be the vmcb02_page allocated by
svm_allocate_nested(), fix vmcb02_page allocation by replacing
alloc_page() with snp_safe_alloc_page().

Signed-off-by: Ashish Kalra <ashish.kalra@amd.com>