*fix certs handling for extended guest requests
Recent refactorings drop InvalidCertificateLength handling for cases
per-instance certificate blobs are in use.

snp_certs_lock is no longer used after this patch, so remove it as well.
It didn't seem to serve a purpose previously since certificate updates
would not affect the results of the firmware requests, and there was no
lock being taken on the writer/update side anyway, so the existing mutex
for issuing requests should be sufficient to maintain previous behavior.

Signed-off-by: Michael Roth <michael.roth@amd.com>