iommu/amd: Setup v2 domain with max pasids
An IOMMU domain can contain multiple devices with different PASID limits.
However, the maximum number of PASIDs of a domain is bounded by the limit
of the IOMMU hardware. This limit is used to determine the number levels
of the GCR3 table.

Therefore, to simplify management of GCR3 table, allocate the table
with the hardware maximum PASID limit by default (value -1) unless
specified otherwise.

Signed-off-by: Vasant Hegde <vasant.hegde@amd.com>