x86/mce/amd: Skip quirky banks during CPU init
In some cases, AMD Thresholding blocks should not be enabled. This could
be due to a known erratum or other issues.

Currently, the hardware registers for the quirky blocks are modified so
that init code will skip enabling the blocks. Register modification was
needed because the registers were read in two passes: first during CPU
init and second during late init.

However, the thresholding blocks are now configured only during CPU
init. So modifying the registers is no longer necessary.

Furthermore, the current quirks apply to all blocks within a bank.
Individual block handling is not needed.

Add a helper function to check for, and skip, quirky banks during CPU
init.

Signed-off-by: Yazen Ghannam <yazen.ghannam@amd.com>