x86/mce/amd: Save bank types only during init
Scalable MCA systems provide values in the MCA_IPID register to identify
the type of an MCA bank. This information is collected as part of
general MCA handling, and it does not need to be cached.

Currently, the bank type information is cached, per-CPU, during CPU
init. The only use for the cached info is to provide unique kobject
names for the thresholding sysfs interface.

Define a helper function and data structure to cache the needed
information only during CPU init. This information will be used to
generate the kobject names, and then the information will be discarded.

Signed-off-by: Yazen Ghannam <yazen.ghannam@amd.com>