x86/mce/amd: Clean up interrupt_enable_store()
A number of tasks are done in threshold_restart_bank(). The tasks are
independent of each other, and they should be done in individual helper
functions. One of these tasks is interrupt enable/disable from sysfs.

Define a helper function to do interrupt enable/disable from sysfs. And
remove an unnecessary "interrupt capable" check. The "interrupt_enable"
attribute is only visible when a block is "interrupt capable".

Signed-off-by: Yazen Ghannam <yazen.ghannam@amd.com>