iommu/amd: Add support for vIOMMU translation device
The AMD IOMMU driver needs to setup a v1 table to store GPA->SPA
translation for each hardware-assisted vIOMMU instance, which will be used
for virtualizing Command buffers, Event log, PPR log, and Completion-wait
Semaphore. The v1 table needs to be setup using an unused DTE, which will
refer to as translated DTE in this patch.

This v1 table in the translated DTE should be the same as the v1 table
of the VFIO domain created when attaching a pass-through device.

TODO: Need to finda a better way to implement support for trans DTE.

Signed-off-by: Suravee Suthikulpanit <suravee.suthikulpanit@amd.com>