perf: core/x86: Add support to register a new vector for PMI handling
Create a new vector in the host IDT for PMI handling wihin a passthrough
PMU implementation. In addition, add a function to allow the registration
of the handler and a function to switch the PMI handler.

This is the preparation work to support KVM passthrough vPMU to handle its
own PMIs without interference from PMI handler of the host PMU.

Signed-off-by: Xiong Zhang <xiong.y.zhang@intel.com>
Signed-off-by: Mingwei Zhang <mizhang@google.com>