KVM: SVM: Enable IBS virtualization on SVM/SEV/SEV-ES guests
To enable IBS virtualization capability on SVM/SEV guests, bit 2 at
offset 0xb8 in VMCB is set to 1 for SVM/SEV guests.

To enable IBS virtualization capability on SEV-ES guests, bit 12 in
SEV_FEATURES in VMSA is set to 1 for SEV-ES guests.

To use the passthrough pmu capability for IBS, set
enable_passthrough_pmu module parameter when IBS virtualization is
enabled on the guest.

Signed-off-by: Manali Shukla <manali.shukla@amd.com>