iommu/amd: Add Support for IOMMU Command Buffer Virtualization w/o GSTBufferTRPMode
In this case, when the guest sets up the guest command buffer
by programming the GPA of the buffer to the MMIO Offset 0008h
(i.e. Command Buffer Base Address Register), the hypervisor traps
the MMIO write, extract the GPA/HVA of the buffer, and pins and map
the pages in the IOMMU v1 table with the IOMMU Private Address (IPA).

Signed-off-by: Suravee Suthikulpanit <suravee.suthikulpanit@amd.com>