x86/sev: ensure that RMP table fixups are reserved for memblock
To handle cases where the RMP table placement in BIOS is not 2M aligned and causes
subsequent kexec failures, the RMP table start and end physical range aligned to
2MB is reserved in all the three e820 tables.

But the reservation is done too late and the remaining memory after the RMP table
in the 2MB range gets allocated to memblock. As this memory range has been
reserved in the e820 tables it is marked as HV-Fixed. With this memory also have been
allocated to memblock, it can potentially be allocated as guest memory which then
causes RMPUPDATE failures during page state transitions as this memory has been marked
as HV-Fixed.

The issue captured below with memblock=debug:

[    0.000000] SEV-SNP: *** DEBUG: snp_probe_rmptable_info:352 - rmp_base=0x280d4800000, rmp_end=0x28357efffff
...
[    0.000000] BIOS-provided physical RAM map:
...
[    0.000000] BIOS-e820: [mem 0x00000280d4800000-0x0000028357efffff] reserved
[    0.000000] BIOS-e820: [mem 0x0000028357f00000-0x0000028357ffffff] usable
...
...
[    0.183593] memblock add: [0x0000028357f00000-0x0000028357ffffff] e820__memblock_setup+0x74/0xb0
...
[    0.203179] MEMBLOCK configuration:
[    0.207057]  memory size = 0x0000027d0d194000 reserved size = 0x0000000009ed2c00
[    0.215299]  memory.cnt  = 0xb
...
[    0.311192]  memory[0x9]     [0x0000028357f00000-0x0000028357ffffff], 0x0000000000100000 bytes flags: 0x0
...
...
[    0.419110] SEV-SNP: Reserving start/end of RMP table on a 2MB boundary [0x0000028357e00000]
[    0.428514] e820: update [mem 0x28357e00000-0x28357ffffff] usable ==> reserved
[    0.428517] e820: update [mem 0x28357e00000-0x28357ffffff] usable ==> reserved
[    0.428520] e820: update [mem 0x28357e00000-0x28357ffffff] usable ==> reserved
...
...
[    5.604051] MEMBLOCK configuration:
[    5.607922]  memory size = 0x0000027d0d194000 reserved size = 0x0000000011faae02
[    5.616163]  memory.cnt  = 0xe
...
[    5.754525]  memory[0xc]     [0x0000028357f00000-0x0000028357ffffff], 0x0000000000100000 bytes on node 0 flags: 0x0
...
...
[   10.080295] Early memory node ranges[   10.168065]
...
node   0: [mem 0x0000028357f00000-0x0000028357ffffff]
...
...
[ 8149.348948] SEV-SNP: RMPUPDATE failed for PFN 28357f7c, pg_level: 1, ret: 2

As seen above, memblock allocation shows the last 1MB of RMP end getting allocated for memblock allocator which is what
the rmptable fixup has made reserved, but memblock allocation has happened before that, so now this memory is usable for
allocations even after being made reserved. This memory range subsequently gets allocated as guest memory and then
causes the RMPUPDATE failure.

This can be potentially fixed by not reserving this memory range in e820_table, but that causes KEXEX_FILE_LOAD syscall
failures during kexec boot as KEXEC_FILE_LOAD syscall uses the in-kernel e820 table, therefore mark this memory not
usable for memblock allocations by doing memblock_reserve().

Though, the RMP table fixups for kexec patch is in 6.9.0-rc7, not marking this fix for stable, as this issue can only
be triggered if SNP is enabled and SNP guests are launched.

Fixes: 400fea4b9651 ("x86/sev: Add callback to apply RMP table fixups for kexec")

Suggested-by: Thomas Lendacky <thomas.lendacky@amd.com>
Reviewed-by: Tom Lendacky <thomas.lendacky@amd.com>
Signed-off-by: Ashish Kalra <ashish.kalra@amd.com>
Signed-off-by: Michael Roth <michael.roth@amd.com>